requirements.txt
Django==1.3.1
PIL==1.1.7
South==0.7.3
cmsplugin-filer==0.8.0
django-classy-tags==0.3.4.1
django-cms==2.2
django-filer==0.8.5
django-mptt==0.5.2
django-photologue==2.3
django-sekizai==0.5
easy-thumbnails==1.0-alpha-20
html5lib==0.90
sorl-thumbnail==11.12
textile==2.1.5
wsgiref==0.1.2
